"Updated .gitignore, backend config, controllers, routes, and server files with new code and formatting changes"

+++ backend/controllers/product.controller.js
@@ -1,7 +1,7 @@
 import Product from "../models/product.model.js";
-
 import mongoose from "mongoose";
 
+// Get all products
 export const getProduct = async (req, res) => {
 	try {
 		const products = await Product.find({});
@@ -12,6 +12,7 @@ export const getProduct = async (req, res) => {
 	}
 };
 
+// Create a new product
 export const createProduct = async (req, res) => {
 	const product = req.body;
 
@@ -38,6 +39,7 @@ export const createProduct = async (req, res) => {
 	}
 };
 
+// Update an existing product
 export const updateProduct = async (req, res) => {
 	const { id } = req.params;
 	const product = req.body;
@@ -60,14 +62,20 @@ export const updateProduct = async (req, res) => {
 	}
 };
 
+// Delete a product
 export const deleteProduct = async (req, res) => {
 	const { id } = req.params;
 
+	if (!mongoose.Types.ObjectId.isValid(id)) {
+		return res
+			.status(404)
+			.json({ success: false, message: "Invalid product id" });
+	}
 	try {
 		await Product.findByIdAndDelete(id);
 		res.status(200).json({ success: true, message: "Product removed" });
 	} catch (error) {
 		console.error("Error in Fetch Products:", error.message);
-		res.status(404).json({ success: false, message: "Product not found" });
+		res.status(500).json({ success: false, message: "Server Error" });
 	}
 };
